Rename personActions to personCardActions
[#152463922]
